[PLAT-12483]Upgrade YBC client and server version to 2.1.0.0-b2: Fixes [PLAT-12442][PLAT-12458]
Summary:
Upgrade YBC server and client version. Fixes include the following YBC commits:
- https://github.com/yugabyte/ybc/commit/ded98e35270e82ff1f6de4034e8a41a00ac54d8a - [PLAT-12458]
```
ISSUE
Currently we have a single timeout (yb_api_timeout_secs) that governs both rpc calls to ybdb and waiting for async yb tasks like create snapshot, restore snapshot etc to complete. We have on occasion seen that we ned to use a very large timeout (1 hour plus) for tasks like restore_snapshot to complete.

FIX
We will use 2 different timeout. We will use yb_api_timeout_secs (default set to 60 ) for rpc call timeouts and we will use a new gflag yb_task_timeout_secs (default set to 600) to wait for async tasks likes create snapshot and restore_snapshot to complete. In addition remove most of the default timeout parameters for methods. This will ensure we are using the timeouts set using the gflags.

TESTING DONE
All unit tests passed. We have done one round of testing with the a db that has 36K tables in it.
```
- https://github.com/yugabyte/ybc/commit/65a79d65c22d0eedfee768bd13d0eb90abce5766 - [PLAT-12442]
```
ISSUE
The line variable is overwritten everytime but we noted a case where the previous request was getting processed.

Fix the bogus compiler warning message due to https://gcc.gnu.org/bugzilla/show_bug.cgi?id=89549.

FIX
Blank out the line variable in the coprocess beofre reading the next message. Also added -Wno-misleading-indentation to the CMakefile to avoid https://gcc.gnu.org/bugzilla/show_bug.cgi?id=89549

TESTING DONE
All unit tests passed
```
- https://github.com/yugabyte/ybc/commit/aa713fc8e162b0a44f3ffcc688ef462c067eb379 - [PLAT-12442]
```
ISSUE
Currently we just keep the coprocess running all of the time. This has 2 issues

We have an additional process running all the time.

We are seeing issues at multiple customers where after running correctly for a while we see the coprocess getting older requests and we need to restart the coprocess to rectify the situation. We typically see this only post number of backups being run. The suspect is that there is a bug in how boost:async_write sends the buffer to stdout of the coprocess or there is some issue where the bash read function ends up reading the previous message again. Given that both of these are not in our hands it is better to just start and stop the coprocess.

FIX
We have a mechanism to start the tablet splitter thread when the first backup/restore is started and to stop when the last concurrent backup/restore is complete. We can just leverage this to start and stop the coprocess too in those methods.

TESTING DONE:
Ran single backup and verified that coprocess starts only when backup is run and stops once backup stops
Ran parallel backups and verified that coprocess starts only when first backup starts and is stopped once all backups were stopped.

Complete unit tests were run and all have passed.
```

Test Plan: dev-itests

Reviewers: vpatibandla, kkg

Reviewed By: kkg

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32104