[PLAT-8500]: IP information and health check label on each nodes is overlapping
Summary:
IP information and health check label on each nodes is overlapping

On health tab we have health checks for each node in each div. We can collapse the box to minimise it from UI. For each node we can see its IP and number of pass/failed checks. In this particular case we have 3 node K8
universe. So while showing the information the label utilizes the complete line passes the information to next line.

Fix: Ensure we show some part of label followed by a tooltip and ... to indicate it is overflowing

Test Plan:
Please refer to the screenshot

{F146720}

Reviewers: jmak

Reviewed By: jmak

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32119