[#20858] DocDB: Simplify Master RegisterPathHandler
Summary:
A dozen or so master web server url paths are handled only on the leader, and follower redirect the request to the leaders.
This is currently done with the use of multiple binds, and placeholders like below:

```
  cb = std::bind(&MasterPathHandlers::HandleTablePageJSON, this, _1, _2);
  server->RegisterPathHandler(
      "/api/v1/table", "Table Info",
      std::bind(&MasterPathHandlers::CallIfLeaderOrPrintRedirect, this, _1, _2, cb), false, false);
```

When dozens of these lines are written right next to each other it is easy to make mistakes.

Helper methods `RegisterPathHandler`, `RegisterLeaderOrRedirect` and `RegisterLeaderOrRedirectWithArgs` help simplify the calling code like below:

```
  RegisterLeaderOrRedirect(
      server, "/api/v1/table", "Table Info", &MasterPathHandlers::HandleTablePageJSON);
```

Renamed `HandleCatalogManager` to `HandleAllTables` and `HandleCatalogManagerJSON` to `HandleAllTablesJSON`

Fixes #20858
Jira: DB-9846

Test Plan: Jenkins tests

Reviewers: tnayak, xCluster, slingam, bogdan

Reviewed By: tnayak

Subscribers: esheng, ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32082