[PLAT-9909]Add Preflight check and selective table restore in "Advanced restore"
Summary: Implemented New Advanced restore modal. this advanced restore modal is written as a wrapper over the already exiting "Restore modal". We do this my modifying the context. So, the logic remains same , but the ui differs and input differs. we map the new input to the old inputs.

Test Plan: Tested manually on multiple scenarios

Reviewers: lsangappa, jmak, rmadhavan

Reviewed By: jmak

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D31511