[#20892] YSQL: Fix divide by 0 error in new cost model when pathtarget width is 0
Summary:
Some queries cause divide by zero error in the following conditions,
* Base scans cost model is enabled (yb_enable_base_scans_cost_model = ON)
* yb_fetch_size_limit is enforced, which happens in following condition where
  result_width is the estimated width of result from docDB to pggate.
  ```if (yb_fetch_size_limit > 0 &&
         (yb_fetch_row_limit < 0 ||
          result_width * yb_fetch_row_limit > yb_fetch_size_limit))```
* result_width = 0, which happens in cases where no columns are projected from
  the base table or aggregate function is pushed down.
  eg. `SELECT 0 FROM test`
      `SELECT count(*) FROM test`

This is a temporary fix for this issue. If result_width = 0, it is set to a
fixed size, which represents the width of the result when the base table has no
primary key and YBCTID is the UUID. The width of the result is 33 in this case.
Jira: DB-9877

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressPlanner'

Reviewers: tverona, tnayak

Reviewed By: tnayak

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D32150