[PLAT-9970][Minor UI] Editing scheduled backups doesn’t population the retention information.
Summary:
when a backup with incremental backup is edited, we calculate the retention period by subtracting the expiryTime with the lastIncrementalBackupTime.
but , we made a mistake by not parsing the "lastIncrementalBackupTime", which resulted an NAN value.

Fix:
parse the "lastIncrementalBackupTime"

Test Plan: Tested manually by setting and checking the "Retention time"

Reviewers: jmak, lsangappa, rmadhavan

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32161