[#20856] YSQL: Fix table rewrite on non-colocated tables/matviews in a colocated DB
Summary:
When re-creating the DocDB table as part of table rewrite, the new table is created with the
default colocation setting for the database itself. Fix this to instead use the colocation relopt
of the original table.

Additionally, add a work-around for GH issue #20914 in `xcluster_ysql_test_base.cc`.
Jira: DB-9843

Test Plan: `yb_matview` and `yb_alter_table`

Reviewers: dsrinivasan, xCluster, hsunder

Reviewed By: dsrinivasan

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D32127