[pg15-upgrade] test refactor: extract common upgrade test code
Summary:
New upgrade tests are showing a lot of duplication, so pull the common code out. It also helps readability, and should encourage more upgrade tests.

Additionally, update the timeout for initdb to work with Linux release.

Additionally, improve/update some of the comments on the various steps.

Test Plan:
On MacOS 13 arm64:
    ./yb_build.sh release --sj
    pg15_tests/test_lint.sh
    pg15_tests/test_upgrade.sh

Reviewers: jason

Reviewed By: jason

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D32162