Replace C parse_type() function with PL/pgSQL
The C code cause a number of challenges, including the requirement for a
superuser to install the extension, and would require a fair bit of
refactoring to abide by [secure design principals]. It also makes
installation difficult on Windows, and would likely be rejected by
organizations like AWS that tend to balk at C code.

So delete the `parse_type()` function and all the C code added in
4ec32e7 and rewrite the `format_type_string()` function based on the
prototype that @ewie developed in the comments of #315. By looking up
the type output format function in the catalog, we're able to correctly
output the data type with its full SQL name as before. The exception is
interval types, which require the PostgreSQL grammar to parse the
interval fields (`second`, `interval minute`, etc.) into a bitmask for
the `typmod`. So require that they be specified exactly as PostgreSQL
outputs them.

Make adjustments to the patches to eliminate hunk offset messages due to
the changes to the SQL code, and upgrade `actions/checkout` to v4 to
eliminate warnings in the workflows.

  [secure design principals]: https://www.postgresql.org/docs/current/extend-extensions.html#EXTEND-EXTENSIONS-SECURITY