[BACKPORT 2.18][PLAT-12445] Fixing permission issue with pexlock
Summary:
Original commit: 4aaabbe387ef1f8f7a8b9025dc6ea5f0c3e5f596 / D32177

Issue sometimes when pexlock is owned by different user we can not delete it. The trap is too
aggressive and sometimes errors in this case. We should do our best to delete but if the file exists
we are usually fine, that means the venv was already created. Whenever the venv needs to be
recreated hopefully the lock file gets cleaned up or customer must manually perform the action.

There has also been discussion whether /tmp is the right place for this lock file. As it's only to
prevent very rapid concurrent attempts to activate_pex (like in itest) there's not a huge concern
with where it lives. If there is desire we can also change the location of the lockfile to someplace
else so this code is safer moving forward. Locks in the new location would always be cleaned up with
the trap and the old lock is irrelevant now, also avoiding permission issues.

Test Plan: activate pex with lockfile owned by root and ensure success.

Reviewers: nsingh, svarshney

Reviewed By: svarshney

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32137