[#20491] Fix CDCSDK Flaky test: CDCSDKYsqlTest.InsertBeforeAfterSnapshot
Summary:
In tsan runs for **CDCSDKYsqlTest.InsertBeforeAfterSnapshot**, the test is failing with assertion
error of not getting WAL records for the insert operations done leading to empty number of records
in GetCDCChanges rpc call.

The reason of this is the write operation over DB spwans aysnc thread and does not wait for the
insert to be completely persisted causing potentail race condition of reading unwritten data that
could cause errors in tsan mode UT runs.

This fix ensures that the write is persisted by reading if the updated value is written or not in a timeout
of 60 seconds.
Jira: DB-9496

Test Plan: Jenkins: test regex: .*CDCSDK*

Reviewers: skumar, asrinivasan

Reviewed By: asrinivasan

Subscribers: ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D31854