[BACKPORT 2.20][PLAT-12554] Use DR api in Repair DR modal and drop table from DR modal and allow empty dbs field for full restart
Summary:
In the repair DR config modal, the option to restart the current DR config sends a restart xCluster config request rather than a restart DR config request.
As we now have dedicated DR config APIs for restarting, this diff replaces the restart xCluster config request with a restart DR config request.
In the drop table modal for DR, we were using the xCluster edit table api instead of the set table api for DR. This diff addresses this by switching to the
DR api when the replication table component is used in DR context.

For the restart DR config request, we did not accept empty `dbs` fields prior to this change.
This means the client must make additional GET requests to map the DR config's table UUIDs to namespace UUIDs on the source universe.
When the only available option is a complete restart of the entire config, the additional network requests are unnecessary.
This diff removes the requirement of non-empty `dbs` field to make the restart DR config request consistent with the restart xCluster config request.
If the user provides an empty `dbs` field, then we will interpret it as a request to restart for all tables in the config.

Original commit: abcdf7c4b120e7151b63d8d6c14f811b591062e0 / D32202

Test Plan:
Verify that the the repair DR modal sends a restart DR config request.
Verify that a restart DR config request with an empty `dbs` field is not rejected.

Reviewers: cwang, hzare, sanketh, #yba-api-review

Reviewed By: cwang, sanketh, #yba-api-review

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32272