[BACKPORT 2.20][#20827] YSQL: Fix inaccurate detection of constants when computing distinct prefix
Summary:
Original commit: 9161aec99413bcfe9c85e035d497f2fe8652290b / D32062
Given that there are two tables t1 and t2 where k is the first column of the primary key in both tables. Consider a query such as

```
SELECT DISTINCT t2.k FROM t1 JOIN t2 ON t1.k = t2.k;
```

When the planner chooses a distinct index scan on t2, we previously erroneously treated t2.k as a constant since an index clause is of the form t2.k = <some terms not related to t2>. However, this behavior is inaccurate since there can be more than one value of t2.k in batch nested loop joins. We now restrict the optimization to scenarios where `t2.k = C` is an index clause and C is equivalent to a "true" constant (i.e. EC_MUST_BE_REDUNDANT in postgres terms).

----------------------------------------------------------------

Example scenario where we can treat t2.k as a constant (but we do not after this change)

Consider a nested loop join such as

```
foreach $1 in t1.k do;
  distinct index scan on t2 where t2.k = $1
```

Here, the distinct index scan can assume that t2.k is a constant for the duration of that scan and hence can exclude t2.k from the distinct prefix.

----------------------------------------------------------------

Example scenario where it is inaccurate to treat t2.k as a constant

Consider a batch nested loop join such as

```
foreach ($1, ..., $1024) in t1.k do;
  distinct index scan on t2 where t2.k in ($1, ..., $1024)
```

Here, a single distinct index scan on t2 can return multiple values for column k since it can be any value from the batch. Thus, we cannot exclude t2.k from the distinct prefix.

Finally, this behavior is not expected in hash or merge joins since this condition needs to be an index condition. However, the join condition is not an index condition on hash or merge joins.
Jira: DB-9816

Test Plan:
Jenkins
./yb_build.sh --java-test TestPgRegressDistinctPushdown

Reviewers: tnayak, smishra, mtakahara

Reviewed By: mtakahara

Subscribers: yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32237