[PLAT-12413] Fix swaggerGen after moving unit test to WithServer
Summary:
As part of https://phorge.dev.yugabyte.com/D32172, we moved the unit tests to start a fake application by extending Play's WithServer so that we can get the application's port used for construting the V2 API path. Previously this was extending WithApplication that did not expose the port. Here the materializer was initialized as part of @Before method.
However, as part of this change, the `swaggerGen` was broken since it starts a play server in a standalone mode (not unit test mode). So the @Before method does not run and the materializer is not initialized.
Fixing this by explicitly using the running application's materializer.

Test Plan: sbt swaggerGen

Reviewers: nsingh, amalyshev, svarshney

Reviewed By: nsingh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32242