[PLAT-12553] Set useHostVPC for providers configured before 2.16
Summary:
GCP providers configured prior to 2.16 does not have `useHostVPC` set even if they are
using the same. Now, when we attempt to update these provider (once they have been used for universe
deployment) - the update operation was failing as we can't edit the same for in-use providers.
This diff fixes the same

Test Plan:
Created provider in 2.16
Deployed universe with it.
Verified editing provider is successful on master

Reviewers: yshchetinin

Reviewed By: yshchetinin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32211