[#20978] DocDB: Disable Wait-on-Conflict in 2.21.0
Summary:
We've discovered the root cause of #18394 is a launch-blocking bug for wait-on-conflict. In particular, the wait queue assumes that only one request will be sent per session to a given tablet. When this assumption is broken, we outright reject requests.

This revision disables the feature by default.
Jira: DB-9954

Test Plan: Jenkins: compile-only

Reviewers: rthallam, bkolagani

Reviewed By: rthallam, bkolagani

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D32257