[#13156] Clear pending deletes on failed delete task
Summary:
Master has pending_deletes_ state tracked on the ts descriptor to track tablets with an outstanding async delete task. This state affects operations the LB can do on certain tablets and tservers. Currently, we only clear from this state when the task is successful. This means that on failures, we incorrectly retain tablets in pending_deletes_.

Fix this by also clearing from pending_deletes_ when the task failed.

There is also another race condition where catalog manager tries to remove from pending_deletes_ before adding to the set. This can also cause tablets to be retained in the set indefinitely after the task has finished. We often see the warning: `Pending delete for tablet ... doesn't exist`. To fix this, we ensure that we always add to the pending deletes set before creating the async task.

Test Plan: Jenkins

Reviewers: asrivastava, mhaddad

Reviewed By: asrivastava

Subscribers: ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32258