[#20667] DocDB: Turn on tablet guardrail mechanism
Summary:
As title.  The value being used for tablet_replicas_per_gib_limit is
from experiments using linear regression determine how the per tablet
overhead varies as a number of tablets is changed.

Depends on D31824
Fixes #20667
Jira: DB-9663

Test Plan:
Temporarily hardwire the RAM total to 6 GB:
```
~/code/yugabyte-db/src/yb/tserver/server_main_util.cc:94:
  total_ram = 6_GB;
```
Then start a cluster:
```
 bin/yb-ctl start  --master_flags 'vmodule=*tablet_limits*=1'
```
and verify that the feature is on with the proper limit:
```
I0207 10:23:40.862541 15181 tablet_limits.cc:128] vlog1: Approved an additional 8 tablet replicas, which will increase the total running tablet replica count to 8, which is still below the safe system maximum of 2963
```

Reviewers: hsunder

Reviewed By: hsunder

Subscribers: ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32260