[BACKPORT 2.18][#20396] YSQL: Fix signaling postmaster death to child backends in Linux environment
Summary:
Original commit: 8c5bd92d8aa4237f7be48eaf6e19ceb7b9d0335b / D31823
Linux provides `PDEATH_SIG` as a mechanism for signaling child processes when their parent process has exited.
In a YugabyteDB environment running on Linux, this needs to be configured on all PG backends soon after they fork from the postmaster process.

Prior to this revision, the `PDEATH_SIG` was configured only after validating `IsYugaByteEnabled()` on the backend.
This function checks for the presence of the pggate object to validate that the process was indeed forked in the context of YugabyteDB.
Since this check was performed soon after the backend fork, pggate was not initialized as yet, and `PDEATH_SIG` never got configured.
This revision uses the `YBIsEnabledInPostgresEnvVar()` instead which checks for the presence of the `YB_ENABLED_IN_POSTGRES` environment variable
to ensure that `PDEATH_SIG` is always configured in a Linux environment.

The macro wrapper for this function call is also altered to directly check for the presence of the header file which is vanilla Postgres' preferred validation mechanism.
Jira: DB-9386

Test Plan:
```
./yb_build.sh --cxx-test pgwrapper_pg_libpq-test --gtest-filter 'PgPostmasterExitTest.SignalBackendOnPostmasterDeath'
./yb_build.sh --cxx-test pgwrapper_pg_libpq-test --gtest-filter 'PgPostmasterExitTest.SignalIdleBackendOnPostmasterDeath'
```

Reviewers: telgersma, amartsinchyk, mbautin, smishra

Reviewed By: telgersma

Subscribers: yql, smishra, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32083