[BACKPORT 2.20][PLAT-12553] Set useHostVPC for providers configured before 2.16
Summary:
GCP providers configured prior to 2.16 does not have `useHostVPC` set even if they are
using the same. Now, when we attempt to update these provider (once they have been used for universe
deployment) - the update operation was failing as we can't edit the same for in-use providers.
This diff fixes the same

Depends on - D32211
Original Commit -
https://github.com/yugabyte/yugabyte-db/commit/ae2d2a75d71155fae03a482f74a1f2a62b41d8b2

Test Plan:
Created provider in 2.16
Deployed universe with it.
Verified editing provider is successful on master

Reviewers: yshchetinin, vbansal

Reviewed By: vbansal

Subscribers: yugaware

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32256