[pg15-upgrade] Add support for upgrading tables with dropped columns
Summary:
When PG drops a column, it maintains an entry for the column in the pg_attribute catalog table with an attname like "........pg.dropped.2........". When a column gets dropped in YB, we maintain the same PG catalog entry, but the YB master's catalog manager, which has its own parallel data structure for the schema, removes the entry for the dropped column in its data structure.

During a ysql major version upgrade from PG11 to PG15, ysql_dump/pg_restore replay the actions of creating the table and dropping the column, to end up with the same dropped column semantics in the new major version's PG catalog. This is the same as the upstream PG mechanism, except that to support online upgrade, in YugabyteDB both catalogs coexist. Prior to this fix, the replay would break the YB catalog manager, because after the original column drop in PG11, when the catalog manager saw the drop column statement coming from pg_restore, it couldn't find the column in the catalog manager's data structure.

Modify catalog manager when in ysql major upgrade mode to:
 1) Allow for dropping a column that doesn't exist.
 2) For debugging purposes, log when the CREATE TABLE request has a different set of columns from the existing catalog manager schema.

This change exposes a problem that didn't manifest during the upgrade previously. As background, in creating the PG15 metadata for each table, pg_restore sends a single string with a number of SQL statements. In the dropped column case they're 1) various "set next oid" and "set next relfilenode" statements, 2) the CREATE TABLE statement, 3) UPDATE the dropped column's pg_attribute table entry, and 4) ALTER TABLE ... DROP COLUMN. PG treats multiple statements sent in one string as executing within a single transaction. Since YugabyteDB doesn't currently support mixing DDLs with modifications to the PG catalog in a single transaction, make a change that causes the ysql_dump/pg_restore process to emit each statement as a separate string, so that they're executed in separate transactions.

As a result of the fix for separate SQL statements, an #ifdef'd out section of dumpTableSchema() in pg_dump.c can be restored. It doesn't have any effect on YB but removing the #ifdef will be less noise when reading or merging.

Test Plan:
Jenkins: rebase: pg15-upgrade

    ./yb_build.sh release --sj
    pg15_tests/test_lint.sh
    pg15_tests/test_upgrade_dropped_col.sh
    pg15_tests/test_upgrade.sh

Reviewers: jason, fizaa, hsunder

Reviewed By: jason, fizaa, hsunder

Subscribers: ybase, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32088