[#18399] DocDB: Display locks used in conflict resolution as opposed to shared in-memory locks for waiters in pg_locks
Summary:
`pg_locks` displays `hash_cols/range_cols` that the transaction locks/wishes to lock. In the existing implementation, `pg_locks` uses doc paths corresponding to the shared in-memory locks (awaiting) for waiters. In certain cases, this can vary slightly from the actual locks/intents being used as part of conflict resolution.

For instance, the statement `UPDATE foo SET b=2 where a=1;` should show that the locks were requested on column `b`. But the doc paths used for determining the shared locks don't use the actual column id, and instead request an in-memory lock on the `kLivenessColumn` (column id 0). Displaying locks corresponding to column id `0` as part of `pg_locks` might be misleading. Additionally, as part of conflict resolution, we check for intents corresponding to column `b` and not `kLivenessColumn`.

This diff alters the behavior of `pg_locks` for waiting transactions by displaying the locks used in conflict resolution instead (which correspond to the actual columns).

Test Plan: ./yb_build.sh --cxx-test='TEST_F(PgGetLockStatusTest, TestWaiterLockContainingColumnId) {'

Reviewers: rsami, tvesely

Reviewed By: rsami

Subscribers: yql, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D31516