[PLAT-12579] Handle failed customerConfig request on configs page
Summary:
**Context**
In StorageConfiguration.js we're filtering the customer config data without checking that the customer
config data is defined first:
```
    const backupListData = customerConfigs.data.filter((list) => {
      if (activeTab === list.name.toLowerCase()) {
        return list;
      }
      return null;
    });
```
If customerConfigs.data is null or undefined, then the UI will not load.

This was introduced in 7d4a5325bf1af0b11fa04f5a6266a4278d34cbf6 and is present in YBA versions 2.8+

**Changes**
Check if the promise state of customerConfigs is error and display an error indicate when applicable.
Fallback to an empty array if customerConfigs.data is undefined/null.

Test Plan:
Block the customer config network request or otherwise cause the request to fail.
Verify that the backup configs page shows an error indicator which informs the user about the failure to fetch
customer configs.
{F148817}

Reviewers: rmadhavan, kkannan, lsangappa, cwang, sanketh

Reviewed By: kkannan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32293