[PLAT-12538] Force bootstrap for repair after failover
Summary:
While repairing DR config after a failover, we should make sure to perform a full bootstrap otherwise in the rare case that bootstrap is skipped, the older primary may still retain changes that are no longer present on the new primary.

This diff also renames the alert template description for new ysql tables.

Test Plan:
1. Create univs A, B
2. Create a new db and a table on both sides with no data
3. Setup DR A -> B
4. Pause underlying xcluster replication
5. Write a row to table on A
6. Failover to B
7. Repair DR config to end up in B -> A
8. Verify that A does not retain the row in step 5.

Prior to this change, A did retain the row in this case.

Reviewers: cwang, jmak, #yba-api-review!

Reviewed By: cwang, jmak

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32210