[#21008] DocDB: Lightweight profiling of callsites
Summary:
We have found that condition variable notifications (signal/broadcast) are sometimes slow.  In order to better understand the issue, in this diff we are adding a lightweight mechanism for counting and timing function call sites by simply wrapping them in the YB_PROFILE macro. This mechanism is easier to use than metrics and allows to identify hot spots in code that is normally expected to be fast, i.e. does not involve waiting on a resource.

Flags:
* `enable_callsite_profile` (false by default) to turn the profiling on;
* `enable_callsite_profile_timing` (false by default) to allow using monotonic time in addition to CPU cycles (reported in microseconds). This might have a higher performance impact but is sometimes necessary.

Web UI endpoint: /pprof/callsite_profile

Specify `?reset=1` to reset all counters to 0.
Jira: DB-9981

Test Plan:
Jenkins

Manual testing on a local cluster started with yb-ctl and running a yb-sample-app test

Run a TPCC perf test on a cluster
(some screenshots of the UI page: https://github.com/yugabyte/yugabyte-db/issues/21008#issuecomment-1936906004)

Reviewers: sergei, xCluster, hsunder

Reviewed By: hsunder

Subscribers: ybase, esheng, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32273