Merged PR 1271766: Improve  performance of the $lt operator
### Does this PR have any customer impact?
Yes
### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Perf
### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No
### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))
Improve $lt performance when having mixed types
### Description
In scenarios where $lt was applied to a field that had "nulls" in it, we would start with MinKey and scan up - which would require passing through all the null valued entries.

Do a best effort to pick a lower bound that is better than MinKey when available.