[PLAT-12473] Add DR error state to YBA UI
Summary:
**Context**
We've added a new `Error` state for disaster recovery (DR) configs with the changes in:
D32334
If the create DR config or replace replica operation fails, YBA will set the DR config to `Error` state.
For DR configs in `Error` state, the user can only delete the DR config. All other operations (including all operations on the underlying xCluster config) are disabled.

**Change**
This diff adds support for the DR config error state to the YBA UI.
We added an optional `drConfigState` parameter to the `getEnabledConfigActions` helper function so that xCluster config's associated with DR can return the current
enabled actions array given the current DR config state.

Test Plan:
Create a DR config between universe A and universe B.
Stop the VM for universe B while the creation task is occurring.
Verify that the DR config is in error state by checking the response of the DR config
GET request and also the config status label in the YBA UI.

Verify that only delete DR config action is allowed.
{F149501}

Reviewers: cwang, sanketh

Reviewed By: sanketh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32338