Run qualified typmodin function in format_type_string()
Formatting typmodin_func with %I causes a qualified function name to be
formatted as a identifier when the namespace is not on the search path.
For example, foo.bar() will be formatted as "foo.bar"() when foo is not
on the search_path.  Calling "foo.bar"() triggers the exception handler
in most cases either because the function doesn't exist or because it's
not a typmodin function.  format_type_string() will then return NULL
which in turn results in col_type_is() to fail with an error that the
wanted type does not exist.

Fix this by formatting the function name with %s.  This works because %s
emits qualified names and quoted identifiers if necessary for regproc
(or any other OID).