[#20417] YSQL: Reduce sleep time in TestPgWriteRestart
Summary:
After 4556775, TestPgWriteRestart fails when run with Fail-on-Conflict
concurrency control mode. This is because the exponential backoff
parameters have been reduced and the total time of query layer retries
for the conflicting transaction is less than the 5000ms sleep in the
blocking transaction.

To fix this, the sleep has been reduced to 500ms (this also reduces
the test time). Moreover, the explicit setting of query layer retries
to 20 has been removed, since the test doesn't depend on it for
reliably testing the retry semantics on conflict errors.
Jira: DB-9411

Test Plan: ./yb_build.sh --java-test org.yb.pgsql.TestPgWriteRestart

Reviewers: bkolagani

Reviewed By: bkolagani

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D32361