[#19934] YSQL: Add a guc to limit the number of rows in pg_locks
Summary:
When a large transaction modifies a large number of rows, the number of
results in pg_locks could result in the system running out of memory.
For example, if we were to perform the following insert:

```
BEGIN;
INSERT INTO t(k, v) SELECT i, i from generate_series(1, 1000000) i;
```

This commit introduces a new GUC `yb_locks_txn_locks_per_tablet` to limit the number
of rows returned by pg_locks. When the GUC is > 0 the rows returned will
be limited, but when it is set to 0 we return all results.
Jira: DB-8892

Test Plan:
./yb_build.sh --cxx-test pg_wrapper-test --gtest_filter PgWrapperFlagsTest.VerifyGFlagDefaults
./yb_build.sh --cxx-test pg_wrapper-test --gtest_filter PgWrapperFlagsTest.TestGFlagOverrides

Reviewers: pjain, tfoucher, bkolagani

Reviewed By: tfoucher, bkolagani

Subscribers: ybase, yql, smishra, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32373